ENVSETUP: Update _stoe_config_read() to manage multiple var definition
It may be possible to define var with multiple appends, so we need to take
care of the multiple lines in _stoe_config_read().
Note that a same var maybe defined in local.conf and site.conf files.
In such case, bitbake always prefers the one from local.conf file.
So apply same filter to display the one set by bitbake.

Change-Id: Id4409fa4c2d51feb9839cf505c4e502057e62d2d
Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>