ENVSETUP: manage Ubuntu required packages check and auto-update
Add new file Ubuntu that list all required packages mentioned in Yocto
mega-manual for both Thud And Dunfell:
  https://www.yoctoproject.org/docs/2.6.4/mega-manual/mega-manual.html#required-packages-for-the-build-host
  https://www.yoctoproject.org/docs/3.1/mega-manual/mega-manual.html#required-packages-for-the-build-host

The file contains:
  <Essentials Packages list> needed to build an image on a headless system>
  <Graphical and Eclipse Plug-In Extras Packages list>
  <OpenEmbedded Self-Test (oe-selftest) Packages list>
Only the <Documentation> packages are not listed, as not needed on our side.

For each supported distribution we manage separated list.

A specific option is also added to allow automatic package update.

Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>
Change-Id: I9daabab9368cb3fd869146030c8da8ea3a479cd0